EntityTypeAlreadyExistsException extend from DataIntegrityViolationException Move EntityTypeAlreadyExistsException to molgenis-data-validation
