+++ molgenis-data-validation/src/main/resources/l10n/data_validation_en.properties
@@ -12,6 +12,8 @@ V10=Cannot delete entity ''{0}'' because entity ''{1}'' depends on it.
 V11=Value '{2}' for attribute '{1}' is referenced by entity '{0}'.
 # Tag validation messages
 V20=Unknown relation IRI ''{1}'' for tag ''{0}''
+# Various validation messages
+V21=Duplicate entity type ''{0}''
 # Package validation messages
 V30=Modifying system package ''{0}'' is not allowed
 V31=Package ''{0}'': identifier ''{1}'' contains illegal characters.
