+++ molgenis-data-validation/src/main/resources/l10n/data_validation_nl.properties
@@ -1,1 +1,2 @@
 V01=Onbekende xref waarde ''{0}'' voor attribuut ''{1}'' van entiteitsoort ''{2}''
+V21=Gedupliceerde entiteitsoort ''{0}''
