+++ molgenis-data/src/main/resources/l10n/data_en.properties
@@ -1,6 +1,5 @@
 D01=Unknown entity type ''{0}''
 D02=Unknown entity ''{0}'' of type ''{1,label}''
-D03=Duplicate entity type ''{0}''
 D04=Unknown attribute ''{1}'' of entity type ''{0,label}''
 D05=Unknown repository ''{0}''
 D06=Unknown repository collection ''{0}''
