+++ molgenis-data/src/main/resources/l10n/data_nl.properties
@@ -1,3 +1,2 @@
 D01=Onbekende entiteitsoort ''{0}''
 D02=Onbekende entiteit ''{0}'' van soort ''{1}''
-D03=Gedupliceerde entiteitsoort ''{0}''
